requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic<2
python-dotenv
openai==0.28.1